Group notebooks together, add bigger discrim version of wgan
The 5stride one works pretty well actually with lower batch sizes, but I
think it'll get stuck where it is because the discriminator actually
thinks the fake images are roughly as fake as the real ones. I'm going
to kick off a pralell experiment with a bigger discriminator hoping the
gap between real/fake will be bigger.